backlight: support setting requested brightness in one step
On some platforms (e.g. ZII RDU1) setting backlight brightness is
relatively slow operations (100s of milliseconds). In this case, attempt
to smoothly change brightness causes operation to last for several
seconds, which is undesirable.

This patch adds 'singlestep' flag to backlight device. If this flag
is set, loop is skipped and requested backlight brightness is set
directly.

Signed-off-by: Nikita Yushchenko <nikita.yoush@cogentembedded.com>