video: backlight: add slew time parameter
Currently the backlight implementation stretches a brightness change over
a period of 100ms. While this is a fine default for PWM backlights, a user
might wish to change this slew time to meet other constraints or even
completely disable it for some backlight devices.

Add a parameter and provide the default value from the backlight device.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>