bbu: add flag to skip writing the padding before the flash header
This is ugly, but needed as on the ZII RDU1 board the flash holds some
important data in the first 1024 bytes, so we need to avoid writing
anything into this. This is implemented by having the barebox partition
start at this offset, but BBU needs to truncate the image accordingly.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>